requirements.txt
Django==5.2.6
postgres==4.0
psycopg2-binary==2.9.10
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-silk==5.4.3
